Gestionando despliegues en Microsoft Fabric, de la centralización a la descentralización/demo/cicd/requirements.txt
azure-identity>=1.15.0
fabric-cicd>=0.1.0